+++ src/Attributes/Route.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rammewerk\Router\Attributes;
+
+use Attribute;
+
+#[Attribute(Attribute::TARGET_CLASS | Attribute::TARGET_METHOD)]
+class Route {
+
+    public function __construct(public string $path = '') {}
+
+
+}